Fixed the type conversion issue

+++ fs/metrics/duration/fs.go
@@ -62,5 +62,5 @@ func Register() {
 
 // SinceInMilliseconds gets the time since the specified start in microseconds.
 func SinceInMilliseconds(start time.Time) float64 {
-	return time.Since(start).Milliseconds()
+	return float64(time.Since(start).Milliseconds())
 }
